+++ fr_persons.php
@@ -27,7 +27,7 @@ if(!isset($_SESSION["fr_person_id"])){ // if "user" not set,
 		</div>
 		<div class="row">
 			<p>
-				<a href="fr_per_create.php" class="btn btn-primary">Add New Volunteer</a>
+				<a href="fr_per_create.php" class="btn btn-primary">Add Volunteer</a>
 				<a href="logout.php" class="btn btn-warning">Logout</a> &nbsp;&nbsp;&nbsp;
 				<a href="fr_persons.php">Volunteers</a> &nbsp;
 				<a href="fr_events.php">Events</a> &nbsp;
@@ -37,7 +37,7 @@ if(!isset($_SESSION["fr_person_id"])){ // if "user" not set,
 			<table class="table table-striped table-bordered" style="background-color: lightgrey !important";>
 				<thead>
 					<tr>
-						<th>Name (Title) (Assignments)</th>
+						<th>Name</th>
 						<th>Email</th>
 						<th>Mobile</th>
 						<th>Action</th>
@@ -57,12 +57,18 @@ if(!isset($_SESSION["fr_person_id"])){ // if "user" not set,
 							echo '<td>'. $row['email'] . '</td>';
 							echo '<td>'. $row['mobile'] . '</td>';
 							echo '<td width=250>';
-							echo '<a class="btn" href="fr_per_read.php?id='.$row['id'].'">Details</a>';
-							echo '&nbsp;';
-							echo '<a class="btn btn-success" href="fr_per_update.php?id='.$row['id'].'">Update</a>';
-							echo '&nbsp;';
-							echo '<a class="btn btn-danger" href="fr_per_delete.php?id='.$row['id'].'">Delete</a>';
-							if($_SESSION["fr_person_id"] == $row['id']) echo " &nbsp;&nbsp;Me";
+							# always allow read
+							echo '<a class="btn" href="fr_per_read.php?id='.$row['id'].'">Details</a>&nbsp;';
+							# only admins can update
+							if ($_SESSION['fr_person_title']=='Administrator')
+								# || $_SESSION['fr_person_id']==$row['id'])
+								echo '<a class="btn btn-success" href="fr_per_update.php?id='.$row['id'].'">Update</a>&nbsp;';
+							# only admins can delete
+							if ($_SESSION['fr_person_title']=='Administrator' 
+								&& $row['countAssigns']==0)
+								echo '<a class="btn btn-danger" href="fr_per_delete.php?id='.$row['id'].'">Delete</a>';
+							if($_SESSION["fr_person_id"] == $row['id']) 
+								echo " &nbsp;&nbsp;Me";
 							echo '</td>';
 							echo '</tr>';
 						}
